Integrate physics & graphics with GameObject
add_*_component() felt awkward. There was nothing preventing you from
adding the same thing as a component on multiple objects, and removing
the component was weird.

The only time I didn't attach sprites / physics bodies to GameObjects
was for stuff that persisted for the entirety of the game, like map
boundaries, the asteroid counter and the FPS display. On the other
hand, it was very easy to forget to attach these to GameObjects which
could have caused weird bugs.